chore: rename files to their class definitions
